upload latest version

+++ api/delete_trip.php
@@ -5,7 +5,7 @@
     user_id, trip_id
   */
 
-include '../credentials/credentials.php';
+  include '../credentials/credentials.php';
   $conn = new mysqli($server_name, $db_username, $db_password, $db_name);
 
   if ($conn->connect_error) {
@@ -13,7 +13,17 @@ include '../credentials/credentials.php';
   }
 
   $trip_id = $_POST['tripID'];
-//   $trip_id = $_POST['trip_id'];
+  $sql = "SELECT ImgUrl FROM Trips WHERE TripID = ".$trip_id;
+  $result = $conn->query($sql);
+  if (!$result) {
+    echo 'Could not run query: ' . mysql_error();
+    exit;
+  }
+  $img_url = $result->fetch_assoc()['ImgUrl'];
+  if(strpos($img_url, "default") === false) {
+    unlink($img_url);
+  }
+
   $sql = "DELETE FROM Trips WHERE TripID = ".$trip_id;
   $conn->query($sql);
   $sql = "DELETE FROM PublicTrips WHERE TripID = ".$trip_id;
@@ -30,11 +40,13 @@ include '../credentials/credentials.php';
   $conn->query($sql);
   $sql = "DELETE FROM Applies WHERE TripID=".$trip_id;
   $conn->query($sql);
-
+  $sql = "DELETE FROM At WHERE TripID=".$trip_id;
+  $conn->query($sql);
   $resp = [
     'status' => 'success',
+    'message' => 'Delete trip succeeded!'
   ];
   $conn->close();
-  header('Content-Type: application/json');
+  // header('Content-Type: application/json');
   echo json_encode($resp);
 ?>
